// src/domain/artifact-set/types.ts
export type ArtifactSet = {
    id: number;
    name: string;
    iconFlowerUrl: string; // <== est aussi l'icone de set.
    iconPlumeUrl: string;
    iconCircletUrl: string;
    iconSandUrl: string;
    iconGobletUrl: string;
    rarityMax: number;
    bonus2P: string;
    bonus4P: string;
    createdAt: string;
};